// types.ts
export enum Currency {
  PEN = 'PEN',
  USD = 'USD'
}

export enum Bank {
  BCP = 'BCP',
  IBK = 'Interbank'
}

export interface Expense {
  id: string;
  name: string;
  category: string; // Ahora es un string dinámico
  date: string;
  periodMonth: string;
  periodYear: string;
  currency: Currency;
  amount: number;
  bank: Bank;
  isAutoGenerated?: boolean;
}

export const DEFAULT_CATEGORIES = [
  'Alimentación', 'Transporte', 'Servicios', 'Entretenimiento', 'Salud', 'Shopping', 'Vivienda', 'Educación', 'Otros'
];